+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+- Fixed active proposals update worker config
+- Fixed saving/updating proposal in the repository
+
 ## [0.2.0] - 2023-07-16
 
 ### Added